.MD option now correctly tags live streams
The .MD option was programmed to output the duration of each video, but this feature
assumed that all videos had a definite duration between 0 seconds and 59:59. This
fails for live streams, which report a duration of -1, so the MD exporter incorrectly
labelled this with (-1:59).

To counter this, .MD now labels live streams with the indication (LIVE) instead of
a time, and prints no indication if the video is not a live stream, but the duration
is a negative number. It also labels correctly videos over an hour, and labels
videos over one day as d:hh:mm:ss.

The MU8 format was not affected by this bug, as the duration must always be specified,
and -1 is the standard value for "not applicable".