Touch up SQL query
Adding an explicit list to avoid a collision between streams.thumbnail_url and
playlists.thumbnail_url.

The downside to this method is that new values that NewPipe may add in the future
are not added automatically to the JSON dump.

Also corrected the description of the variable PlaylistDir at getPlaylists()